improve importing modules, added validation

+++ ClassDefinitions.py
@@ -5,85 +5,13 @@ import warnings                      #użyłam aby wyłączyć ostrzeżenia spow
 import os                            #do czyszczenia ekranu
 import msvcrt                       #do pobierania znaku z klawiatury
 
+
 #zmiana koloru czcionki
 def ChangeColorToPurple(text):
     print('\033[95m' + text + '\033[0m')
 
 def ChangeColorToWhite(text):
     print('\033[0m'+ text + '\033[0m')
-
-#główna funkcja menu
-def menu():
-    object = Product("",0) #tworze przykładowy obiekt na rzecz którego bedę wywoływać metody (funkcje w klasie), "" oznacza ze nazwa domyslnie jest pusta
-    #0 oznacza ze cena domyslnie po utworzeniu to 0, przewiń sobie do definicji placy Product i tam w pierwszych liniach masz _init_ coś tam
-    #za pomocą tego konstruktora możesz sobie tworzyć obiekt z wartościami podanymi w nawiasie
-    
-    zm = 1 #zmienna obslugujaca przeskakiwanie
-    while(1): #petla nieskonczona, warunek zawsze jest spelniony, dlatego za kazym razem jak wcisniesz coś to program bedzie aktualizowal sie
-        os.system('cls') #czyszcze ekranpo kliknieciu każdego znaku
-        
-        print("------------------------------------------------------------------------------------")
-        print("|                               SYSTEM OBSLUGI MAGAZYNU                            |")
-        print("------------------------------------------------------------------------------------")
-        if zm == 1:
-            ChangeColorToPurple("\nDODAJ PRODUKT")#jesli zmiana jest rowna 1 czyli wybrano pozycje 1 to zmien kolor 
-        else:
-            ChangeColorToWhite("\nDODAJ PRODUKT")#jak nie to wypisz ten tekst normalnie w kolorze bialym
-        if zm==2:
-            ChangeColorToPurple("MODYFIKUJ PRODUKT")
-        else:
-            ChangeColorToWhite("MODYFIKUJ PRODUKT")
-        if zm==3:
-            ChangeColorToPurple("USUN PRODUKT")
-        else:
-            ChangeColorToWhite("USUN PRODUKT")
-        if zm==4:
-            ChangeColorToPurple("FILTROWANIE PRODUKTOW")
-        else:
-            ChangeColorToWhite("FILTROWANIE PRODUKTOW")
-        if zm==5:
-            ChangeColorToPurple("WYSWIETL WSZYTSKIE PRODUKTY")
-        else:
-            ChangeColorToWhite("WYSWIETL WSZYTSKIE PRODUKTY")
-        if zm==6:
-            ChangeColorToPurple("WYJSCIE")
-        else:
-            ChangeColorToWhite("WYJSCIE")
-
-        znak = msvcrt.getch() #ta sunkcja zwraca kod znaku, ktory jest wcisniety
-        
-        if znak == b'H': #jesli znak to strzalka w gore to idziemy ze zmianą w góre
-            zm -= 1
-        if znak == b'P': #jesli strzalka w dol to zmiekszamy zmienna i idziemy na dól
-            zm += 1
-        if zm == 7: #jak dojdziemy do końca bo mamy 7 pozycji to wracamy na początek czyli ustawiomy zmiane na 1 i lecimy dalej
-            zm = 1
-
-        
-        if (znak == b'\r') and (zm == 6): #jesli znak to enter (b'\r' - taki ma kod) i zmiana jest 6 to wylazimy z programu
-            os.system("cls" if os.name == "nt" else "clear")
-            close_connection() #zamykam polaczenie z baza
-            exit(1) #koniec programu
-
-        if (znak == b'\r') and (zm == 1): #jak jest enter i zmiana to 1 czyli wybrano 1 pozycje to idzie do classdefinitions do funkcji add_product
-            object.add_product()
-
-        if (znak == b'\r') and (zm == 3): #tu tak samo i niżej tez
-            object.update_product()
-        
-        if (znak == b'\r') and (zm == 4):
-            filter_products()
-        
-        if (znak == b'\r') and (zm == 2):
-            object.update_product()
-        
-        if (znak == b'\r') and (zm == 5):
-            sort_products()
-
-
-
-###############################################################################################################
-
 
 class Product:
     def __init__(self, name, price): 
@@ -93,7 +21,7 @@ class Product:
             self.name = name
             self.price = price
 
-    def add_product(self):
+    def add_product(self,menu):
         #tu tak samo, to sie wykona na produkcie na ktorym to wywolujesz 
         #czyli jak wywolalas object.add-product to sie tak jakby dane tego object zaktualizuja
         os.system('cls')
@@ -105,13 +33,23 @@ class Product:
         
         cursor = Connection.conn.cursor()
         try:
-            cursor.execute(f"INSERT INTO Products (name, price) VALUES ('{self.name}', '{self.price}')") 
-            #cursor jest z pliku connection i on pozwala wykonywać coś w tej bazie czyli on tak jakby trzyma wszytskie dane o tym polączeniu
-            # i np pozwala ci czytac cos z bazy, pobierac jakies wyniki
-            #kwerenda z SQL czyli bezposrednio do bazy to leci i to sie w niej wykonuje bo baza ma swoj specjalny jezyk
-            #do tabeli Products w kolumne name proce dodaj wartosci nazwa i cena ktore sa wprowadzone z klawiatury
-            Connection.conn.commit()
-            print("\nRekord został dodany do bazy danych.")
+            #tu bede sprawdzac czy przypadkiem nie ma juz produktu o takiejsamej nazwie juz
+            #pierwsze wybieram sobie z bazy i ładuje do cursora wszytskie elementy ktore mają taką sama nazwe jak wpisałam z klawiatury
+            cursor.execute(f"SELECT * FROM Products WHERE name LIKE '{self.name}'")#cursr ma teraz w sobie to co wybierze zapytanie
+            wiersze = cursor.fetchall()  
+            #teraz dziele cobie zapytanie na wierszee pojedyncze bo normalnie to taki ciąg znaków
+            #jeżeli dlugosc (rozmiar) tego wyniku jest = 0, to znaczy ze zapytanie nic nie wybralo to znaczy ze nie ma takiego produktu
+            if len(wiersze) != 0:
+                print("\nW bazie istnieje już produkt o takiej nazwie. ")
+            else:
+                #jeżeli nie ma takiego produktu to możesz dodac do bazy a jak jest to wyswietl komunikat i wyjscie
+                cursor.execute(f"INSERT INTO Products (name, price) VALUES ('{self.name}', '{self.price}')")
+                #cursor jest z pliku connection i to taki jakby kontener na dane, on pozwala wykonywać coś w tej bazie czyli on tak jakby trzyma wszytskie dane o tym polączeniu
+                # i np pozwala ci czytac cos z bazy, pobierac jakies wyniki
+                #kwerenda z SQL czyli bezposrednio do bazy to leci i to sie w niej wykonuje bo baza ma swoj specjalny jezyk
+                #do tabeli Products w kolumne name proce dodaj wartosci nazwa i cena ktore sa wprowadzone z klawiatury
+                Connection.conn.commit()
+                print("\nRekord został dodany do bazy danych.")
         except Exception as e:
             print(f"\nBlad podczas dodawania rekordu: {e}")#jak system wykryje blad i sie nie doda poprawnie to wywala bląd
     
@@ -122,7 +60,7 @@ class Product:
 
 
     #tu jest prawie to samo
-    def delete_product(self):
+    def delete_product(self,menu):
         os.system('cls')
         print("------------------------------------------------------------------------------------")
         print("|                               USUWANIE PRODUKTU                                  |")
@@ -131,11 +69,22 @@ class Product:
         
         cursor = Connection.conn.cursor()
         try:
-            cursor.execute(f"DELETE FROM Products WHERE name LIKE '{self.name}'")
-            Connection.conn.commit()
-            print("\nUsuneto rekordy z bazy danych!")
+            #tu znowu sprawdzam czy wgl taki produkt co wpisalam istnieje w bazie
+            #posluguje sie prawie takim samym zapytaniem jak wczesniej, w sensie ze wybieram sobie z bazy produkty o takiej nazwie
+            #jezeli zapytanie zwroci 0  wynikow to znaczy ze nie ma takiego produktu i siema
+            cursor.execute(f"SELECT * FROM Products WHERE name LIKE '{self.name}'")
+            wiersze = cursor.fetchall()  
+            
+            if len(wiersze) == 0:
+                print("\nW bazie nie istnieje produkt o takiej nazwie. ")
+            else:
+                #jesli istnieje to usun go takim zapytaniem z DELETE
+                cursor.execute(f"DELETE FROM Products WHERE name LIKE '{self.name}'")
+                Connection.conn.commit()
+                print("\nUsuneto rekordy z bazy danych!")
         except Exception as e:
             print(f"\nBlad podczas usuwania rekordow: {e}")
+       
         ChangeColorToPurple("\nWYJSCIE")
         msvcrt.getch()
         menu()
@@ -143,65 +92,41 @@ class Product:
 
         #tu jest troche wiecej zabawy bo jak wchodzisz do trybu aktualizowania to mozesz wybrac co chcesz zaktualizowac czyli masz tak jakby drugie menu
         #ktore dziala tak samo tylko zamiast zm jest choice zeby zmienne sie nie dublowały
-    def update_product(self):
+    def update_product(self, menu):
         choice = 1
-        while(1):
-            os.system('cls')
-            print("------------------------------------------------------------------------------------")
-            print("|                               MODYFIKACJA PRODUKTU                                |")
-            print("------------------------------------------------------------------------------------")
-            
-            self.name = input("\nWprowadz nazwe produktu ktory chcesz zmodyfikowac: ")
-            if choice == 1:
-                ChangeColorToPurple("\nZMIEN NAZWE")
-            else:
-                ChangeColorToWhite("\nZMIEN NAZWE")
-            if choice == 2:
-                ChangeColorToPurple("ZMIEN CENE")
-            else:
-                ChangeColorToWhite("ZMIEN CENE")
-
-            zn = msvcrt.getch()
-            
-            if zn == b'H':
-                choice -= 1
-            if zn == b'P':
-                choice += 1
-            if choice == 3:
-                choice = 1
-
-            #jesli wybrałes 1 i enter jest wzisniety to przechodzisz jakby do zmieniania nazwy
-            if (zn == b'\r') and (choice == 1):
-                new_name = input("Wprowadz nowa nazwe produktu: ")
-                cursor = Connection.conn.cursor()
-                try:
-                    cursor.execute(f"UPDATE Products SET name = '{new_name}' WHERE name LIKE '{self.name}'")
-                    #tym zapytaniem aktualizujesz wszytsko w tabeli Products co ma nazwe self.name czyli nazwe tego oiektu na ktorym
-                    #wywolujesz i wprowadzasz nowa nazwe ktarą wpisalaś na klawiaturze
-                    Connection.conn.commit()#wysylanie do bazy danych zapytania
-                    print("\nZaktualizowano rekordy!")#jak sie uda to ci mowi ze jest git a jak nie to ze blad, tak samo jest z aktualizacja ceny
-                    ChangeColorToPurple("\nWYJSCIE")
-                    msvcrt.getch()
-                    menu()
-                except Exception as e:
-                    print(f"\nBlad podczas aktualizowania rekordow {e}")
-                
-            if (zn == b'\r') and (choice == 2):
-                new_price = input("\nWprowadz nowa cene produktu: ")
-                cursor = Connection.conn.cursor()
-                try:
-                    cursor.execute(f"UPDATE Products SET price = '{new_price}' WHERE name LIKE '{self.name}'")
-                    Connection.conn.commit()
-                    print("\nZaktualizowano rekordy!")
-                    ChangeColorToPurple("\nWYJSCIE")
-                    msvcrt.getch()
-                    menu()
-                except Exception as e:
-                    print(f"\nBlad podczas aktualizowania rekordow {e}")
+        os.system('cls')
+        print("------------------------------------------------------------------------------------")
+        print("|                               MODYFIKACJA CENY PRODUKTU                           |")
+        print("------------------------------------------------------------------------------------")
+        #wprowadzam nazwe produktu ktory chce zmodyfikowac
+        self.name = input("\nWprowadz nazwe produktu ktory chcesz zmodyfikowac: ")
+        
+        #tu sprawdzam tak jak wczesiej czy taki produkt wgl jest w bazie 
+        cursor = Connection.conn.cursor()
+        cursor.execute(f"SELECT * FROM Products WHERE name LIKE '{self.name}'")
+        wiersze = cursor.fetchall()  
+        
+        if len(wiersze) == 0:
+            print("\nW bazie nie istnieje produkt o takiej nazwie. ")
+            ChangeColorToPurple("\nWYJSCIE")
+            msvcrt.getch()
+            menu()
+        else:
+            #jesli istnieje to sobie robie modyfikacje ceny
+            new_price = input("\nWprowadz nowa cene produktu: ")
+            try:
+                cursor.execute(f"UPDATE Products SET price = '{new_price}' WHERE name LIKE '{self.name}'")
+                Connection.conn.commit()
+                print("\nZaktualizowano rekordy!")
+                ChangeColorToPurple("\nWYJSCIE")
+                msvcrt.getch()
+                menu()
+            except Exception as e:
+                print(f"\nBlad podczas aktualizowania rekordow {e}")
         
 #tu sie konczą funkcje ktore bedziemy wywolywac na rzecz jednego obiektu czyli dodawanie, ususwnaie, 
 # modyfikacja itp bo teraz bedziemy wyswietlac wszytskie produkty, albo jakas ich czesc w każdym razie nie jakis jeden konkretny
-def sort_products():
+def sort_products(menu):
     choice = 1
     while(1):
         os.system('cls')
@@ -242,25 +167,25 @@ def sort_products():
             #tutaj zamiast wyswietlac wszytsko robie ppunkcje ktora wyswietla wszytskie produkty które wybierze zapytanie
             #wszedzie to wyglada tak samo tylko rożnią sie zapytaniaa
             query = 'SELECT * FROM Products ORDER BY name ASC'
-            display_all_product_by_query(query)
+            display_all_product_by_query(query,menu)
             
         if (zn == b'\r') and (choice == 2):
             query = 'SELECT * FROM Products ORDER BY name DESC'
-            display_all_product_by_query(query)
+            display_all_product_by_query(query,menu)
         
         if (zn == b'\r') and (choice == 3):
             query = 'SELECT * FROM Products ORDER BY price ASC'
-            display_all_product_by_query(query)
+            display_all_product_by_query(query,menu)
         
         if (zn == b'\r') and (choice == 4):
             query = 'SELECT * FROM Products ORDER BY price DESC'
-            display_all_product_by_query(query)
+            display_all_product_by_query(query,menu)
 
 
 #funkcja do wyswietlania produktów które zostały wybrane zaputaniem(query) którebędzie różne, jest używana przy sortowniu i przy filtrowaniu
 #za kazdym razem dziala tak samo tylko ze zmienia sie zapytanie wiec zmieniaja sie dane wyswietlane
 #dodałam to żeby za kazdym razem nie powtarzac tego bo sie kod robi na pół kilometra
-def display_all_product_by_query(query):
+def display_all_product_by_query(query,menu):
     os.system('cls')
     print("------------------------------------------------------------------------------------")
     print("|                          WSZYTSKIE PRODUKTY W MAGAZYNIE                          |")
@@ -275,7 +200,7 @@ def display_all_product_by_query(query):
     menu()
 
 
-def filter_products():
+def filter_products(menu):
     os.system('cls')
     print("------------------------------------------------------------------------------------")
     print("|                         FILTROWANIE PRODUKTOW PO CENIE                            |")
@@ -283,7 +208,7 @@ def filter_products():
     price_min = input("\nWprowadz poczatek przedzialu cenowego: ")
     price_max = input("\nWprowadz koniec przedzialu cenowego: ")
     query = f"SELECT * FROM Products WHERE price BETWEEN '{price_min}' and '{price_max}' ORDER BY price ASC"
-    display_all_product_by_query(query)
+    display_all_product_by_query(query,menu)
 
 
 #zamykanie polaczenia z baza
